hardware/requirements.txt
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.2
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keyboard==0.13.5
MarkupSafe==2.1.5
MouseInfo==0.1.3
outcome==1.3.0.post0
pillow==10.2.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytweening==1.2.0
requests==2.31.0
selenium==4.18.1
simple-websocket==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
